update constructor - separated init in different crate
