--- v0_5/debug/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "delegation_v0_5-debug"
-version = "0.1.3"
-authors = ["andrei-marinica <andrei_m_marinica@yahoo.com>"]
-edition = "2018"
-
-[dependencies]
-elrond-wasm = "0.6.2"
-elrond-wasm-debug = "0.6.2"
-
-# elrond-wasm = { path = "../../elrond-wasm-rs/elrond-wasm" }
-# elrond-wasm-debug = { path = "../../elrond-wasm-rs/elrond-wasm-debug" }
-
-delegation_v0_5 = { path = ".." }
