+++ v0_5/src/lib.rs
@@ -17,104 +17,19 @@ pub mod settings;
 pub mod reset_checkpoints;
 pub mod reset_checkpoint_types;
 
-use node_storage::types::*;
-use crate::events::*;
-use node_storage::node_config::*;
-use crate::rewards::*;
-use crate::node_activation::*;
-use crate::user_stake::*;
-use crate::user_unstake::*;
-use user_fund_storage::user_data::*;
-use user_fund_storage::fund_transf_module::*;
-use user_fund_storage::fund_view_module::*;
-use crate::settings::*;
-use crate::reset_checkpoints::*;
-use elrond_wasm_module_pause::*;
+pub use node_storage::types::*;
+pub use crate::events::*;
+pub use node_storage::node_config::*;
+pub use crate::rewards::*;
+pub use crate::node_activation::*;
+pub use crate::user_stake::*;
+pub use crate::user_unstake::*;
+pub use user_fund_storage::user_data::*;
+pub use user_fund_storage::fund_transf_module::*;
+pub use user_fund_storage::fund_view_module::*;
+pub use crate::settings::*;
+pub use crate::reset_checkpoints::*;
+pub use elrond_wasm_module_pause::*;
 
-imports!();
-
-#[elrond_wasm_derive::contract(DelegationImpl)]
-pub trait Delegation {
-
-    // METADATA
-
-    #[endpoint]
-    fn version(&self) -> &'static str {
-        env!("CARGO_PKG_VERSION")
-    }
-
-    // MODULES
-
-    #[module(EventsModuleImpl)]
-    fn events(&self) -> EventsModuleImpl<T, BigInt, BigUint>;
-
-    #[module(NodeConfigModuleImpl)]
-    fn node_config(&self) -> NodeConfigModuleImpl<T, BigInt, BigUint>;
-
-    #[module(RewardsModuleImpl)]
-    fn rewards(&self) -> RewardsModuleImpl<T, BigInt, BigUint>;
-
-    #[module(ResetCheckpointsModuleImpl)]
-    fn reset_checkpoints(&self) -> ResetCheckpointsModuleImpl<T, BigInt, BigUint>;
-
-    #[module(SettingsModuleImpl)]
-    fn settings(&self) -> SettingsModuleImpl<T, BigInt, BigUint>;
-
-    #[module(PauseModuleImpl)]
-    fn pause(&self) -> PauseModuleImpl<T, BigInt, BigUint>;
-
-    #[module(UserStakeModuleImpl)]
-    fn user_stake(&self) -> UserStakeModuleImpl<T, BigInt, BigUint>;
-
-    #[module(NodeActivationModuleImpl)]
-    fn node_activation(&self) -> NodeActivationModuleImpl<T, BigInt, BigUint>;
-
-    #[module(UserUnStakeModuleImpl)]
-    fn user_unstake(&self) -> UserUnStakeModuleImpl<T, BigInt, BigUint>;
-
-    #[module(UserDataModuleImpl)]
-    fn user_data(&self) -> UserDataModuleImpl<T, BigInt, BigUint>;
-
-    #[module(FundTransformationsModuleImpl)]
-    fn fund_transf_module(&self) -> FundTransformationsModuleImpl<T, BigInt, BigUint>;
-
-    #[module(FundViewModuleImpl)]
-    fn fund_view_module(&self) -> FundViewModuleImpl<T, BigInt, BigUint>;
-    
-
-    // Callbacks can only be declared here for the moment.
-
-    #[callback]
-    fn auction_stake_callback(&self,
-            #[callback_arg] node_ids: Vec<usize>,
-            call_result: AsyncCallResult<VarArgs<BLSStatusMultiArg>>) {
-
-        self.node_activation().auction_stake_callback(
-            node_ids,
-            call_result).unwrap();
-        // TODO: replace unwrap with typical Result handling
-    }
-
-    #[callback]
-    fn auction_unstake_callback(&self,
-            #[callback_arg] node_ids: Vec<usize>,
-            call_result: AsyncCallResult<VarArgs<BLSStatusMultiArg>>) {
-
-        self.node_activation().auction_unstake_callback(
-            node_ids,
-            call_result).unwrap();
-            // TODO: replace unwrap with typical Result handling
-    }
-
-    #[callback]
-    fn auction_unbond_callback(&self,
-            #[callback_arg] node_ids: Vec<usize>,
-            call_result: AsyncCallResult<VarArgs<BLSStatusMultiArg>>) {
-
-        self.node_activation().auction_unbond_callback(
-            node_ids,
-            call_result).unwrap();
-            // TODO: replace unwrap with typical Result handling
-    }
-
-}
+#[macro_use]
+extern crate elrond_wasm;
