+++ v0_5_init/mandos/activate_nodes.scen.json
