+++ v0_5_init/mandos/staking_2.scen.json
