+++ v0_5_init/mandos/steps/activate_nodes.steps.json
@@ -139,7 +139,7 @@
                     "nonce": "*",
                     "balance": "1,000,000,000",
                     "storage": "*",
-                    "code": "file:../../output/delegation.wasm"
+                    "code": "file:../../output/delegation_v0_5_init.wasm"
                 },
                 "''auction_contract______________s1": {
                     "nonce": "0",
