+++ v0_5_init/mandos/steps/stake_1.steps.json
