+++ v0_5_init/mandos/steps/stake_2.steps.json
