+++ v0_5_init/mandos/withdraw_inactive..scen.json
