+++ v0_5_init/wasm/src/lib.rs
@@ -1,5 +1,5 @@
 
 #![no_std]
 
-pub use delegation_v0_5::*;
+pub use delegation_v0_5_init::*;
 pub use elrond_wasm_output::*;
